Break up specs so they are interface focused
This commit breaks up the large `SimpleSpec` into its public module
constituents. This provides a more granular and easier to manage
overview of `network`'s test coverage. More test coverage has also been
added to `Network.Socket.ByteString.Lazy` as it was mostly uncovered.
