+++ network.cabal
@@ -115,10 +115,13 @@ test-suite spec
   default-language: Haskell2010
   hs-source-dirs: tests
   main-is: Spec.hs
-  other-modules: RegressionSpec
-                 SimpleSpec
+  other-modules:
+    Network.Test.Common
+    Network.SocketSpec
+    Network.Socket.ByteStringSpec
+    Network.Socket.ByteString.LazySpec
   type: exitcode-stdio-1.0
-  ghc-options: -Wall -threaded
+  ghc-options: -Wall -Werror -threaded
   -- NB: make sure to versions of hspec and hspec-discover
   --     that work together; easiest way is to constraint
   --     both packages to a small enough version range.
